requirements.txt
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0
dash-bootstrap-components>=1.5.0
